renamed js files and exported style message into a new function
