+++ app/javascript/chat/close_chatroom.js
