+++ app/javascript/chat/fetch_chatroom.js
@@ -1,23 +1,12 @@
-import { closeChat } from "./closeChatroom";
+import { closeChat } from "./close_chatroom";
 import { cleanInputMessage } from "./cleanInput";
+import { styleMessage } from "./style_message";
 
 import consumer from "../channels/consumer";
 import Rails from '@rails/ujs';
 
 
-const styleMessage = () => {
-  const allMessages = document.querySelectorAll('[data-user-id]');
-  const lastMessage = allMessages[allMessages.length - 1]
-  const lastMessageAuthorId = lastMessage.dataset.userId;
-  const currentUserId = document.getElementById('current-user-id').innerText;
 
-  if(lastMessageAuthorId != currentUserId) {
-    console.log(lastMessageAuthorId != currentUserId)
-    lastMessage.firstElementChild.classList.add('friend')
-    lastMessage.lastElementChild.classList.add('friend') 
-    lastMessage.lastElementChild.style.color = '#4C5A5B'
-  }
-}
 
 const fetchChatroom = () => {
   const actions = document.querySelectorAll('.chat-action');
